Update Script_ENG_description.txt
Changed the approach to the delete button, because it turns out that for hidden videos it is not nth-child(3), but nth-child(1). Initially there were options to make a check for nth-child(2) to check for search error, but it's waste of code; or to make a check for words, but it should be done for all major languages, so that the script would be multi-lingual. In the end, digging into the page code, I found that delete is the only line of code generated that contains a breakdown into sub-elements “Remove from playlist” + its name. In the end, the universal solution came by itself - to make a search on the element containing this breakdown. Simple and ingenious at the same time